+++ CHANGELOG.md
@@ -1941,3 +1941,4 @@
 - [2026-06-12 16:15] chore(analytics): update dependencies and cleanup unused imports in analytics
 - [2026-06-12 16:18] chore(navigation): update dependencies and cleanup unused imports in navigation
 - [2026-06-12 17:18] style(perf): harmonize light and dark theme palette for perf
+- [2026-06-12 17:54] fix(cache): fix memory leak by cancelling cache subscription